feat(supabase): add typed wrappers for chart manifest and series RPCs
